Add files via upload
During this hackathon, I had the opportunity to work on an engaging Data Analysis project focusing on Electric Vehicles (EVs). 🚗💨 Exploring a comprehensive dataset of EVs, I leveraged Python's Plotly library to conduct extensive Exploratory Data Analysis (EDA). 📊📈 The goal was to gain valuable insights into the trends and patterns in EVs' adoption and characteristics.

The analysis involved univariate and bivariate visualizations, showcasing the distribution of Electric Vehicle Types, Electric Range, Base MSRP, and more. 📍 Additionally, I crafted an interactive Choropleth map to visualize the geographic distribution of EV vehicles based on location, highlighting EV adoption across different cities and states. 🗺️

One of the most exciting aspects of this project was creating a captivating Racing Bar Plot! 🏎️📊 This animated visualization showcased the evolution of EV Makes and their counts over the years, providing a dynamic look at the fast-paced development of the EV industry. 🚀

It was an enriching experience to delve into the world of data-driven insights and apply Python programming skills to bring the data to life. 🐍💡 I'm grateful to Innomatics Research Labs for organizing this hackathon and providing an excellent platform to sharpen our data analysis and visualization abilities.